requirements.txt
python-socketio==5.11.0
eventlet==0.35.2
